Rename 1.18 package
