+++ .github/workflows/maven.yml
@@ -9,7 +9,7 @@ on:
 
 jobs:
   # Build 1.18.2 NMS
-  v1_18_R2:
+  v1_18:
     runs-on: ubuntu-latest
     steps:
       - name: Set up JDK 17 # 1.18.2 can only be built with Java 17
@@ -25,15 +25,15 @@ jobs:
             ~/.m2/repository/org/spigotmc/spigot/1.18.2-R0.1-SNAPSHOT/
             ~/.m2/repository/org/spigotmc/spigot-parent/
             ~/.m2/repository/org/spigotmc/minecraft-server/
-          key: ${{ runner.os }}-v1_18_R2
-          restore-keys: ${{ runner.os }}-v1_18_R2
+          key: ${{ runner.os }}-v1_18
+          restore-keys: ${{ runner.os }}-v1_18
       - name: Cache Maven packages
         id: cacheMain
         uses: actions/cache@v3
         with:
           path: ~/.m2
-          key: ${{ runner.os }}-m2-v1_18_R2
-          restore-keys: ${{ runner.os }}-m2-v1_18_R2
+          key: ${{ runner.os }}-m2-v1_18
+          restore-keys: ${{ runner.os }}-m2-v1_18
 
       - name: Setup BuildTools
         run: mkdir BuildTools && wget -O BuildTools/BuildTools.jar https://hub.spigotmc.org/jenkins/job/BuildTools/lastSuccessfulBuild/artifact/target/BuildTools.jar
@@ -137,7 +137,7 @@ jobs:
   # Build Movecraft
   build:
     runs-on: ubuntu-latest
-    needs: [v1_18_R2, v1_19_R3, v1_20_R4]
+    needs: [v1_18, v1_19_R3, v1_20_R4]
 
     steps:
       - name: Checkout Movecraft
@@ -160,8 +160,8 @@ jobs:
             ~/.m2/repository/org/spigotmc/spigot/1.18.2-R0.1-SNAPSHOT/
             ~/.m2/repository/org/spigotmc/spigot-parent/
             ~/.m2/repository/org/spigotmc/minecraft-server/
-          key: ${{ runner.os }}-v1_18_R2
-          restore-keys: ${{ runner.os }}-v1_18_R2
+          key: ${{ runner.os }}-v1_18
+          restore-keys: ${{ runner.os }}-v1_18
       - name: Cache 1.19.4 Maven package
         uses: actions/cache@v3
         with:
