+++ modules/v1_18/src/main/java/net/countercraft/movecraft/compat/v1_18/IWorldHandler.java
@@ -1,4 +1,4 @@
-package net.countercraft.movecraft.compat.v1_18_R2;
+package net.countercraft.movecraft.compat.v1_18;
 
 import net.countercraft.movecraft.MovecraftLocation;
 import net.countercraft.movecraft.MovecraftRotation;
@@ -21,7 +21,6 @@ import net.minecraft.world.level.chunk.LevelChunk;
 import net.minecraft.world.level.chunk.LevelChunkSection;
 import net.minecraft.world.ticks.LevelChunkTicks;
 import net.minecraft.world.ticks.ScheduledTick;
-import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.block.data.BlockData;
