+++ modules/v1_18/src/main/java/net/countercraft/movecraft/compat/v1_18/NextTickProvider.java
@@ -1,20 +1,15 @@
-package net.countercraft.movecraft.compat.v1_18_R2;
+package net.countercraft.movecraft.compat.v1_18;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.levelgen.structure.BoundingBox;
 import net.minecraft.world.ticks.LevelChunkTicks;
-import net.minecraft.world.ticks.LevelTicks;
 import net.minecraft.world.ticks.ScheduledTick;
-import org.bukkit.Bukkit;
-import org.bukkit.World;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.lang.reflect.Field;
 import java.util.List;
-import java.util.Queue;
 import java.util.stream.Stream;
 
 public class NextTickProvider {
