+++ pom.xml
@@ -14,7 +14,7 @@
 
     <modules>
         <module>modules/api</module>
-        <module>modules/v1_18_R2</module>
+        <module>modules/v1_18</module>
         <module>modules/v1_19_R3</module>
         <module>modules/v1_20_R4</module>
         <module>modules/datapack</module>
